providerを別フィアルに分離、追加フォームのクリア処理、insert処理の不具合修正
