+++ lib/presentation/screen/list_screen.dart
@@ -6,60 +6,9 @@ import '../../domain/ticket_repository.dart';
 import '../../infrastructure/firebase_auth_adapter.dart';
 import '../login_user.dart';
 import 'detail_screen.dart';
+import 'ticket_provider.dart';
 import 'webview_screen.dart';
 import 'welcome_screen.dart';
-
-class TicketNotifier extends StateNotifier<List<Ticket>> {
-  TicketNotifier() : super([]);
-
-  void getList(List<Ticket> tickets) {
-    state = tickets;
-  }
-
-  void delete(Ticket ticket) {
-    var newItems = [
-      for (final aTicket in state)
-        if (ticket.id != aTicket.id) aTicket,
-    ];
-    state = _updateSort(newItems);
-  }
-
-  void insert(String title) {
-    var newItems = [Ticket(id: "", title: title, body: "", sort: 0), ...state];
-    state = _updateSort(newItems);
-  }
-
-  void reorder(int oldIndex, int newIndex) {
-    if (oldIndex < newIndex) {
-      newIndex -= 1;
-    }
-    var newItems = [...state];
-    final Ticket item = newItems.removeAt(oldIndex);
-    newItems.insert(newIndex, item);
-
-    state = _updateSort(newItems);
-  }
-
-  List<Ticket> _updateSort(List<Ticket> _items) {
-    List<Ticket> newItems = [];
-    int lastIndex = _items.length - 1;
-    _items.asMap().forEach((index, ticket) {
-      int newSort = lastIndex - index;
-      if (ticket.sort == newSort) {
-        return;
-      }
-      newItems.add(ticket.copyWith(sort: newSort));
-    });
-    newItems.sort((a, b) {
-      return a.sort - b.sort;
-    });
-    return [..._items];
-  }
-}
-
-final _itemsProvider =
-    StateNotifierProvider.autoDispose<TicketNotifier, List<Ticket>>(
-        (ref) => TicketNotifier());
 
 class ListScreen extends StatefulHookConsumerWidget {
   static const routeName = "ListScreen";
@@ -82,7 +31,7 @@ class _ListScreenState extends ConsumerState<ListScreen> {
     _titleController = TextEditingController();
     Future(() async {
       ref
-          .read(_itemsProvider.notifier)
+          .read(ticketListProvider.notifier)
           .getList(await TicketRepository().getList());
     });
   }
@@ -93,15 +42,17 @@ class _ListScreenState extends ConsumerState<ListScreen> {
     super.dispose();
   }
 
-  void _onTicketAddTapped() {
+  void _onTicketAddTapped() async {
     if (_titleController.text.isEmpty) {
       return;
     }
 
-    ref.read(_itemsProvider.notifier).insert(_titleController.text);
-    _updateSort(ref.read(_itemsProvider));
-
+    var title = _titleController.text;
     _titleController.text = "";
+    var newTicket = await TicketRepository()
+        .upsert(Ticket(id: "", title: title, body: "", sort: 0));
+    ref.read(ticketListProvider.notifier).insert(newTicket);
+    _updateSort(ref.read(ticketListProvider));
   }
 
   void _onListItemTapped(BuildContext context, int index) async {
@@ -120,15 +71,15 @@ class _ListScreenState extends ConsumerState<ListScreen> {
   void _onListItemDeleteTapped(Ticket ticket) async {
     await TicketRepository().delete(ticket);
 
-    ref.read(_itemsProvider.notifier).delete(ticket);
-    _updateSort(ref.read(_itemsProvider));
+    ref.read(ticketListProvider.notifier).delete(ticket);
+    _updateSort(ref.read(ticketListProvider));
 
     _setStateInitView();
   }
 
   void _onReorder(int oldIndex, int newIndex) {
-    ref.read(_itemsProvider.notifier).reorder(oldIndex, newIndex);
-    _updateSort(ref.read(_itemsProvider));
+    ref.read(ticketListProvider.notifier).reorder(oldIndex, newIndex);
+    _updateSort(ref.read(ticketListProvider));
   }
 
   void _onLogoutTapped() async {
@@ -151,7 +102,7 @@ class _ListScreenState extends ConsumerState<ListScreen> {
 
   @override
   Widget build(BuildContext context) {
-    final _items = ref.watch<List<Ticket>>(_itemsProvider);
+    final _items = ref.watch<List<Ticket>>(ticketListProvider);
 
     return GestureDetector(
         onTap: () => FocusScope.of(context).unfocus(),
