+++ lib/presentation/screen/ticket_provider.dart
@@ -1,0 +1,55 @@
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+
+import '../../domain/ticket.dart';
+
+class TicketNotifier extends StateNotifier<List<Ticket>> {
+  TicketNotifier() : super([]);
+
+  void getList(List<Ticket> tickets) {
+    state = tickets;
+  }
+
+  void delete(Ticket ticket) {
+    var newItems = [
+      for (final aTicket in state)
+        if (ticket.id != aTicket.id) aTicket,
+    ];
+    state = _updateSort(newItems);
+  }
+
+  void insert(Ticket ticket) {
+    var newItems = [ticket, ...state];
+    state = _updateSort(newItems);
+  }
+
+  void reorder(int oldIndex, int newIndex) {
+    if (oldIndex < newIndex) {
+      newIndex -= 1;
+    }
+    var newItems = [...state];
+    final Ticket item = newItems.removeAt(oldIndex);
+    newItems.insert(newIndex, item);
+
+    state = _updateSort(newItems);
+  }
+
+  List<Ticket> _updateSort(List<Ticket> _items) {
+    List<Ticket> newItems = [];
+    int lastIndex = _items.length - 1;
+    _items.asMap().forEach((index, ticket) {
+      int newSort = lastIndex - index;
+      if (ticket.sort == newSort) {
+        return;
+      }
+      newItems.add(ticket.copyWith(sort: newSort));
+    });
+    newItems.sort((a, b) {
+      return a.sort - b.sort;
+    });
+    return [..._items];
+  }
+}
+
+final ticketListProvider =
+    StateNotifierProvider.autoDispose<TicketNotifier, List<Ticket>>(
+        (ref) => TicketNotifier());
